Moved killApp() method to Page class
