+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/SavePointTest.kt
@@ -7,7 +7,6 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
-import org.odk.collect.android.support.CollectHelpers
 import org.odk.collect.android.support.StorageUtils
 import org.odk.collect.android.support.pages.AppClosedPage
 import org.odk.collect.android.support.pages.FormEntryPage
@@ -113,7 +112,7 @@ class SavePointTest {
         rule.setUpProjectAndCopyForm("two-question-audit.xml", listOf("external_data_10.zip"))
             .fillNewForm("two-question-audit.xml", "Two Question")
             .answerQuestion("What is your name?", "Alexei")
-            .let { simulateProcessDeath() }
+            .let { simulateProcessDeath(it) }
 
         // Start blank form and check save point is loaded
         rule.fillNewFormWithSavepoint("two-question-audit.xml")
@@ -154,7 +153,7 @@ class SavePointTest {
         rule.editForm("two-question-audit.xml", "Two Question")
             .clickGoToStart()
             .answerQuestion("What is your name?", "Alexei")
-            .let { simulateProcessDeath() }
+            .let { simulateProcessDeath(it) }
 
         // Edit instance and check save point is loaded
         rule.editFormWithSavepoint("two-question-audit.xml")
@@ -196,7 +195,7 @@ class SavePointTest {
         // Create save point for blank form
         rule.fillNewForm("two-question-audit.xml", "Two Question")
             .answerQuestion("What is your name?", "Alexei")
-            .let { simulateProcessDeath() }
+            .let { simulateProcessDeath(it) }
 
         // Check editing instance doesn't load save point
         rule.editForm("two-question-audit.xml", "Two Question")
@@ -220,7 +219,7 @@ class SavePointTest {
         rule.editForm("two-question-audit.xml", "Two Question")
             .clickGoToStart()
             .answerQuestion("What is your name?", "Alexei")
-            .let { simulateProcessDeath() }
+            .let { simulateProcessDeath(it) }
 
         // Check starting blank form does not load save point
         rule.fillNewForm("two-question-audit.xml", "Two Question")
@@ -237,7 +236,7 @@ class SavePointTest {
     /**
      * Simulate a "process death" case where an app in the background is killed
      */
-    private fun simulateProcessDeath() {
-        CollectHelpers.killApp()
+    private fun simulateProcessDeath(page: FormEntryPage) {
+        page.killApp()
     }
 }
